Updated file structure.
